Create FIRST ASSIGNMENT OF PW
Q1. Explain with an example each when to use a for loop and a while loop.
Q2. Write a python program to print the sum and product of the first 10 natural numbers using for
    and while loop.
Q3. Create a python program to compute the electricity bill for a household.
    The per-unit charges in rupees are as follows: For the first 100 units, the user will be charged Rs. 4.5 per
    unit, for the next 100 units, the user will be charged Rs. 6 per unit, and for the next 100 units, the user will
    be charged Rs. 10 per unit, After 300 units and above the user will be charged Rs. 20 per unit.
    You are required to take the units of electricity consumed in a month from the user as input.
    Your program must pass this test case: when the unit of electricity consumed by the user in a month is
    310, the total electricity bill should be 2250.
Q4. Create a list of numbers from 1 to 100. Use for loop and while loop to calculate the cube of each
    number and if the cube of that number is divisible by 4 or 5 then append that number in a list and print
    that list.
Q5. Write a program to filter count vowels in the below-given string.
    string = "I want to become a data scientist"